requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.6
pandas>=2.0.0
pyarrow>=14.0.0
